Refactored threads to protocol folder
