+++ open-pond-protocol/src/lib.rs
@@ -1,11 +1,9 @@
 //! Library that implements the Open Pond Protocol
-mod app_manager;
 mod config;
 mod message;
-mod peer_pool;
+mod protocol;
 mod queue;
-mod servicer;
 
 pub use crate::config::*;
-pub use crate::peer_pool::*;
-pub use crate::servicer::*;
+pub use crate::protocol::peer_pool::start_peer_pool;
+pub use crate::protocol::servicer::start_servicer;
